refactor(formRichtext): change types to Redactor.d.ts file
