+++ src/components/ui/form-richtext/FormRichtext.vue
@@ -1,6 +1,7 @@
 <script setup lang="ts">
 import { computed, getCurrentInstance, onMounted, onUnmounted } from 'vue';
-import type { FormRichtextProps, RedactorLibraryType } from './types';
+import type { FormRichtextProps } from './types';
+import type { RedactorLibraryType } from './redactor/Redactor';
 import Redactor from './redactor/redactor.usm';
 
 const props = withDefaults(defineProps<FormRichtextProps>(), {
